refactor: small code clean
