+++ src/features/search/pages/Search/ThematicSearch/types.ts
@@ -1,0 +1,6 @@
+import { Offer } from 'shared/offer/types'
+
+export type ThematicSearchPlaylistData = {
+  title: string
+  offers: { hits: Offer[] }
+}
